bmt: Merge Standalone asynchasher to bmt package for easier diff
